added RESTful category controller
